Give every stage a CLI, and fix the defects that were safe to fix
Paths were hardcoded to one developer's machine (D:/AT_Master/..., F:/team project/...),
so no stage ran anywhere else. Each script now takes argparse arguments; sfm_pipeline.py
resolves COLMAP from --colmap, then $COLMAP_EXE, then PATH, and gained --skip-dense and
--no-show so it can run headless.

Defects fixed:

- sfm_pipeline.py had a stray '")' left over from translating its comments to English,
  which made the file unparseable -- the pipeline entry point could not be imported at
  all. The pre-translation revision in the history is clean.
- incremental_pipeline.py formatted time.time()*1000, a float, with the :010d integer
  spec. Only reachable with snapshot_images_freq > 0, so it had never fired.
- match_shi_sift.py truncated to 250 matches before RANSAC and without sorting, so the
  geometric check saw an arbitrary subset in keypoint order. Matches are now sorted by
  descriptor distance and the pre-RANSAC truncation is gone; --num_matches caps the
  drawing only. On frame_95/frame_96 this takes RANSAC's input from 250 to all 559
  candidates, of which 350 are inliers.
- calibrate_fisheye.py used "== None", and let cv2.fisheye.calibrate's bare
  "fabs(norm_u1) > 0" assertion escape. That assertion names neither the offending view
  nor what to do about it; it is now caught and re-raised with the diagnosis procedure,
  and --exclude acts on the answer.

No algorithm or numeric parameter changed, so the archived results still describe the
same computation. The two defects that would change results -- the placeholder camera
and the lossy descriptor round-trip -- are left alone and documented.

+++ 05_reconstruction/incremental_pipeline.py
@@ -37,7 +37,9 @@ def save_reconstruction_snapshot(reconstruction, snapshot_path):
         snapshot_path (pathlib.Path): Folder path to save snapshot.
     """
     logging.info("Creating snapshot...")
-    timestamp = time.time() * 1000
+    # int(): time.time() * 1000 is a float, and the :010d format spec rejects
+    # floats with a ValueError. Only reachable when snapshot_images_freq > 0.
+    timestamp = int(time.time() * 1000)
     path = snapshot_path / f"{timestamp:010d}"
     path.mkdir(exist_ok=True, parents=True)
     logging.verbose(1, f"=> Writing snapshot to: {path}")
